% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFxns.R
\name{validatePopShares}
\alias{validatePopShares}
\title{Validate population shares}
\usage{
validatePopShares(PopShares, tol = 1e-06)
}
\arguments{
\item{PopShares}{A numeric vector of population shares.}

\item{tol}{Tolerance for checking if shares sum to 1.}
}
\value{
TRUE if valid, otherwise throws an error.
}
\description{
Checks that population shares are valid: non-negative and sum to 1.
}
\keyword{internal}
