% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarRep.R
\name{SDRepresentation}
\alias{SDRepresentation}
\title{Compute the amount of representation left unexplained by a random sampling model.}
\usage{
SDRepresentation(PopShares, BodyN, a = -0.5, b = 1, nMonte = 10000)
}
\arguments{
\item{PopShares}{A numeric vector containing the group-level population proportions.}

\item{BodyN}{A positive integer denoting the size of the political body in question.}

\item{a, b}{Parameters controlling the affine transformation for how the representation measure is summarized.
That is, \code{a} and \code{b} control how the expected L1 deviation of the population shares from the body shares
is re-weighted. The expected L1 deviation is the average value of the absolute deviation of the population from body shares under
a random sampling model. This expected L1 deviation is multiplied by \code{a}; \code{b} is as an additive re-scaling term: \code{a*E[L1]+b}.
By default, \code{a=-0.5} and \code{b=1} so that the expected Rose Index of Proportionality is used in the calculation.}

\item{nMonte}{A positive integer denoting number of Monte Carlo iterations used to approximate the variance of representation under a random sampling model.}
}
\value{
A scalar summary of the amount of representation not explained by a random sampling model.
More precisely, this function returns the
the residual standard deviation when using the expected degree of representation to predict observed
representation under a random sampling model.
}
\description{
Finds the residual standard deviation when using the expected representation for any group
in a political body to predict observed representation as described in Gerring, Jerzak and Oncel, 2023.
}
\examples{

SDRep <- SDRepresentation(PopShares = c(1/3, 2/3, 1/3),
                                BodyN = 50)

print( SDRep )

}
